backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5
djoser>=2.2.0
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
psycopg2-binary>=2.9.9
Pillow>=10.1.0
python-decouple>=3.8
redis>=5.0.0
drf-yasg>=1.21.7
djangorestframework-simplejwt[crypto]
dj-database-url>=2.1.0
